feat(cities): add filtered search functionality and consolidate DTOs
- Implement `findAllWithFilters` in `CitiesService` using QueryBuilder to support name searching and limit constraints.
- Add a new GET `/cities/search` endpoint in `CitiesController` to handle filtered city queries.
- Consolidate city-related DTOs into a single `dto/city.dto` file and add `GetCitiesDto` for validation and type safety.
- Update service logic to use case-insensitive matching (ILIKE) for city names.
