--- src/cities/dto/update-city.dto.ts
@@ -1,4 +1,0 @@
-import { PartialType } from '@nestjs/swagger';
-import { CreateCityDto } from './create-city.dto';
-
-export class UpdateCityDto extends PartialType(CreateCityDto) {}
